requirements.txt
langchain==0.3.19
langchain_community==0.3.18
langchain_core==0.3.37
langchain_groq==0.2.4
langchain_huggingface==0.1.2
langchain_mistralai==0.2.7
langchain_qdrant==0.2.0
langgraph==0.2.74
python-dotenv==1.0.1
qdrant_client==1.13.2
SQLAlchemy==2.0.38
typing_extensions==4.12.2
